chore: remove reflection zero type when zero value is known (PROJQUAY-6620)
reflection uses unsafe types and is cpu intensive. we can reduce cpu
cycles on every api call by simple comparing an empty strings equality
to the value of the AuthToken.

strings.EqualFold is a std lib solution that provides similar
functionality, but that is not necessary for comparing empty strings as
it simply uses the standard equality operator as well

Topic: rm-reflection

Relative: projquay-6620
Signed-off-by: Ross Bryan <robryan@redhat.com>